Move all object-related files into evaluation/object
