+++ evaluation/object/constants.go
@@ -1,0 +1,5 @@
+package object
+
+var (
+
+)
